requirements.txt
streamlit==0.82.0
Pillow==8.2.0
azure-cognitiveservices-vision-computervision==0.8.0
msrest==0.6.21

